requirements.txt
watchdog>=3.0.0
pyairtable>=3.1.1
python-dotenv>=1.1.1
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3